unify proof verification
